+++ repositories/client_db_repository.go
@@ -1,0 +1,3 @@
+package repositories
+
+type ClientDbRepository struct{}
